Unify tfrecord_loader and TFRecordDataset
In a similar fashion to the TFRecord writer, which has a single write
function for both `Example` and `SequenceExample`, this makes it so that
`tfrecord_loader` and `TFRecordDataset` can handle both formats and are
triggered based on if the `sequence_description` variable is set or not.

As the `sequence_description` needs to be set to a value that is not
None for the SequenceExample to be triggered a new default has to be set
for if all sequence features should be extracted, as such either an
empty list or dict being passed in would cause all features to be
extracted both for Example and SequenceExample.