Fix a bug where MultiTFRecordDataset reads the records multiple times (#99)
When infinite=False, we would expect the MultiTFRecordDataset to read
each record once only. This was not the case because sharding was not
enabled for MultiTFRecordDataset, so each process would read all files
once.

Enable sharding and add tests to make sure sampling works correctly,
and that we read the dataset only once.